+++ internal/cluster/clusterworkflow/delete_node_pool_label_set_activity.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package workflow
+package clusterworkflow
 
 import (
 	"context"
